rename module normalize_cfg{=>s}
