+++ hugr-passes/src/lib.rs
@@ -18,14 +18,14 @@ pub mod linearize_array;
 pub use linearize_array::LinearizeArrayPass;
 pub mod lower;
 mod monomorphize;
-pub mod normalize_cfg;
+pub mod normalize_cfgs;
 pub mod untuple;
 
-/// Merge basic blocks. Subset of [normalize_cfg], use the latter.
-#[deprecated(note = "Use normalize_cfg", since = "0.15.1")]
+/// Merge basic blocks. Subset of [normalize_cfgs], use the latter.
+#[deprecated(note = "Use normalize_cfgs", since = "0.15.1")]
 pub mod merge_bbs {
     #[expect(deprecated)] // remove this
-    pub use super::normalize_cfg::merge_basic_blocks;
+    pub use super::normalize_cfgs::merge_basic_blocks;
 }
 
 pub use monomorphize::{MonomorphizePass, mangle_name, monomorphize};
